solution for isLIstPalindrome in Python
the recursion solutions time out or use too much memory. my ultimate solution uses a list reversal where I reverse the linkages of the linked list up to halfway.

sadly, most people are changing the linked list to a list despite the problem, which goes against the point of the exercise.